requirements.txt
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.1
httpx==0.24.1
mnemonic==0.20
Pillow==9.5.0
pycparser==2.21
PyNaCl==1.5.0
pypng==0.20220715.0
python-dateutil==2.8.2
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.4.2
ripemd-hash==1.0.0
safe-pysha3==1.0.4
six==1.16.0
techbureau-symbol-sdk-python==3.0.7.dev2
typing_extensions==4.6.2
urllib3==2.0.2
websocket-client==1.5.2
